Protected more routes with login auth

+++ udemy/graphql-prisma/src/resolvers/Mutation.js
@@ -53,27 +53,25 @@ const Mutation = {
       token: jwt.sign({ userId: user.id }, 'thisisasecret')
     };
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    const userExists = await prisma.exists.User({ id: args.id });
-
-    if (!userExists) {
-      throw new Error('User not found');
-    }
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
 
     return prisma.mutation.deleteUser(
       {
         where: {
-          id: args.id
+          id: userId
         }
       },
       info
     );
   },
-  async updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id
+          id: userId
         },
         data: args.data
       },
@@ -81,7 +79,6 @@ const Mutation = {
     );
   },
   createPost(parent, args, { prisma, request }, info) {
-    console.log('REQUEST', request);
     const userId = getUserId(request);
 
     return prisma.mutation.createPost(
@@ -100,7 +97,19 @@ const Mutation = {
       info
     );
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) {
+      throw new Error('Unable to delete post');
+    }
+
     return prisma.mutation.deletePost(
       {
         where: {
